pyproject.toml
[project]
name = "hackaton_model_training"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Alexey Laletin", email = "alex.klark.laletin@yandex.com"},
    {name = "George Kiselev", email = "gsx2002@yandex.ru"}
]
dependencies = [
    "tqdm>=4.66.5",
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "pytest>=8.3.2",
    "python-dotenv>=1.0.1",
    "mlflow>=2.16.0",
    "dvc>=3.55.2",
    "click>=8.1.7",
    "onnx>=1.16.2",
    "onnxruntime>=1.19.2",
    "onnxscript>=0.1.0.dev20240910",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "APACHE 2.0"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
data = [
    "bing-image-downloader>=1.1.2",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
